Separated skin and functionality
Searching and image selection based on category selection now happens in
a separate listener.

Skin now functioning more smoothly with the new selection mode. This is
due to the fact that the Terra skin mouseup and mousedown methods are
not used anymore. However it also removes the checkbox handling, but as
this was not used in our application it seems to be okay to do.